dom/base/test/test_urlExceptions.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=926890
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 926890</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=926890">Mozilla Bug 926890</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe name="x" id="x"></iframe>
  <iframe name="y" id="y"></iframe>
</div>
<pre id="test">
</pre>
  <script type="application/javascript">

  // URL.href throws
  var url = new URL('http://www.example.com');
  ok(url, "URL created");

  var status = false;
  try {
    url.href = '42';
  } catch(e) {
    status = true;
  }
  ok(status, "url.href = 42 should throw");

  url.href = 'http://www.example.org';
  ok(true, "url.href should not throw");

  status = false
  try {
    new URL('42');
  } catch(e) {
    status = true;
  }
  ok(status, "new URL(42) should throw");

  status = false
  try {
    new URL('http://www.example.com', '42');
  } catch(e) {
    status = true;
  }
  ok(status, "new URL(something, 42) should throw");

  </script>
</body>
</html>
